routes_fix

+++ finance-service/internal/service/category_service.go
@@ -3,7 +3,10 @@ package service
 import (
 	"context"
 	"finance-service/internal/repository"
+	"finance-service/internal/utils"
 	"finance-service/proto/pb"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"net/http"
 )
@@ -18,6 +21,10 @@ func NewCategoryService(repo *repository.CategoryRepository) *CategoryService {
 }
 
 func (c *CategoryService) CreateCategory(ctx context.Context, req *pb.CreateCategoryRequest) (*pb.AbsResponse, error) {
+	companyId := utils.GetCompanyId(ctx)
+	if companyId == "" {
+		return nil, status.Error(codes.Aborted, "error while getting company from context")
+	}
 	if err := c.repo.CreateCategory(req.Name, req.Desc); err != nil {
 		return nil, err
 	}
@@ -27,6 +34,10 @@ func (c *CategoryService) CreateCategory(ctx context.Context, req *pb.CreateCate
 	}, nil
 }
 func (c *CategoryService) DeleteCategory(ctx context.Context, req *pb.DeleteAbsRequest) (*pb.AbsResponse, error) {
+	companyId := utils.GetCompanyId(ctx)
+	if companyId == "" {
+		return nil, status.Error(codes.Aborted, "error while getting company from context")
+	}
 	if err := c.repo.DeleteCategory(req.Id); err != nil {
 		return nil, err
 	}
@@ -36,5 +47,9 @@ func (c *CategoryService) DeleteCategory(ctx context.Context, req *pb.DeleteAbsR
 	}, nil
 }
 func (c *CategoryService) GetAllCategory(ctx context.Context, req *emptypb.Empty) (*pb.GetAllCategoryRequest, error) {
+	companyId := utils.GetCompanyId(ctx)
+	if companyId == "" {
+		return nil, status.Error(codes.Aborted, "error while getting company from context")
+	}
 	return c.repo.GetAllCategory()
 }
